+++ packages/bundler-webpack/src/index.ts
@@ -3,3 +3,4 @@ export {
   type WebpackConfig,
 } from "./adapter/create-config.js";
 export { webpackAdapter } from "./adapter/index.js";
+export { webpack } from "./plugin-helper.js";